requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.11.8
colorama==0.4.4
importlib-metadata==2.0.0
iniconfig==1.1.1
iriusrisk-python-client-lib==1.0.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
PyYAML==5.3.1
six==1.15.0
toml==0.10.2
urllib3==1.26.1
zipp==3.4.0
